pyproject.toml
[build-system]
requires = ["uv_build>=0.8.22"]
build-backend = "uv_build"

[project]
name = "aind-behavior-just_frames"
description = "A library for benchmarking online video acquisition/encoding pipelines"
authors = [ {name = "Bruno Cruz", email = "bruno.cruz@alleninstitute.org"}]
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Operating System :: Microsoft :: Windows",
]
version = "0.5.0rc0"
readme = {file = "README.md", content-type = "text/markdown"}

dependencies = [
  "aind_behavior_services>=0.12, <0.13",
  "pydantic-settings"
]

[project.optional-dependencies]

launcher = ["aind-clabe[aind-services] >= 0.8.2 ,<0.9.0"]


[dependency-groups]

dev = ["aind_behavior_just_frames[launcher]",
    "ruff",
    "codespell"
]

docs = [
    'Sphinx',
    'furo',
    'sphinx-jinja',
    'autodoc_pydantic[erdantic]',
    'myst_parser',
    'sphinx-jsonschema',
]

[project.scripts]
just-frames = "aind_behavior_just_frames.cli:main"

[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.ruff.lint]
extend-select = ['Q', 'RUF100', 'C90', 'I']
extend-ignore = []
mccabe = { max-complexity = 14 }
pydocstyle = { convention = 'google' }

[tool.codespell]
skip = '.git,*.pdf,*.svg'
ignore-words-list = 'nd'

[tool.uv]
default-groups = ['dev']
required-version = '>=0.8.4'